restructured MailTemplates API
