+++ src/Accounts/Accounts.tsx
@@ -17,6 +17,7 @@ import {
 } from '@devexpress/dx-react-grid-material-ui';
 import Paper from '@material-ui/core/Paper';
 import React, { useEffect, useState } from 'react';
+import { createAccount, deleteAccount, fetchAccounts, updateAccount } from '../api';
 import {
   Command,
   DefaultColumnsTypeProvider,
@@ -28,13 +29,7 @@ import {
   PopupEditing,
   UsersTypeProvider,
 } from '../common/Table';
-import {
-  createAccount,
-  deleteAccount,
-  fetchAccounts,
-  fetchUserGroups,
-  updateAccount,
-} from '../DataServices/DataServices';
+import { fetchUserGroups } from '../DataServices/DataServices';
 import { UserGroupProps, UserGroupsData } from '../UserGroups/types';
 
 const DEFAULT_COLUMNS = [
